Changed comments and  split tests into smaller classes jenkins-ignore
